Change RPC server from JSON-RPC over websockets to using gRPC
The new design of the RPC server is MUCH cleaner, more efficient
and less buggy. We're ready to have bmclient connect to it.

Removed support for Go 1.3 because gRPC needs Go 1.4. Removed go vet
from goclean.sh. It's proving to be an unnecessary burden since it
doesn't understand grpc.Errorf and always complains. It can be run
manually, if needed.